+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 - unreleased
 
 - Added LookbackHandler
+- Fixed `inc` function used in the StatusHandler
 
 ## 1.0.0 - 2019-03-21